+++ artiq/examples/master/repository/speed_benchmark.py
@@ -102,7 +102,7 @@ class SpeedBenchmark(EnvExperiment):
                               self.scheduler.priority, None, False)
 
     def run_without_scheduler(self, pause):
-        payload = globals()["_Payload" + self.payload](*self.managers())
+        payload = globals()["_Payload" + self.payload](self)
 
         start_time = time.monotonic()
         for i in range(int(self.nruns)):